requirements.txt
Django==5.1.5
django-ninja==1.3.0
PyJWT==2.10.1
django-cors-headers==4.6.0
Pillow==11.1.0
qrcode==8.0
python-decouple==3.8
